+++ src/Pim/Bundle/ImportExportBundle/Transformer/AbstractOrmTransformer.php
@@ -20,6 +20,8 @@ use Pim\Bundle\ImportExportBundle\Transformer\ColumnInfo\ColumnInfoTransformerIn
  * @author    Antoine Guigan <antoine@akeneo.com>
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ *
+ * @abstract
  */
 abstract class AbstractOrmTransformer
 {